Fix aspnet/AspNetCore#6186
The issue here is that the compiler generates a different syntax tree
for `data-` attributes :(. I've added a note to #6112 to follow up on
the compiler behavior. We used to parse data- attributes differently for
components, but I really don't think having the compiler special case
them makes sense.

For now I've added a workaround here to lower `data-` attributes in a
fashion similar to other dynamic attribute values.