Fix a bug with ref and declaration phase
This fixes a regression in the new `@ref` support, where the build will
fail during the declaration pass.

The problem is that the EliminateMethodBodyPass runs really early and
wipes out the usage of `@ref` before we have a chance to see it. The
solution is to make it run really late. This makes sense because we
basically just use this pass to do cleanup.