Don't let background IOExceptions and UnauthorizedAccessException land in ActivityLog.
- This reduces the noise when a user renames a .razor file in VS.
- Added tests to validate that we properly swallow IOExceptions and UnauthorizedAccessException.

During the fix of this I found:
1. It doesn't crash VS
2. It's unavoidable because our project manager can't guarantee that an underlying file won't disappear before processing it.
3. In VS' case renaming a Razor file doesn't result in the project system detecting its removal until after we've attempted to parse the deleted file.
4. This exception is reported to the activity log as an error.

I took the approach of fixing 4 because this is a pretty common case that can't always be avoided. I wish I could mitigate it by dequeuing background parsing work; however, we aren't notified that it's actually gone until we've already thrown an exception.

aspnet/AspNetCore#11720