Improve DefaultLSPEditorfeatureDetector.IsLSPEditorFeatureEnabled performance. (#1893)
Most of the cost in this method is currently due to VSFeatureflags.IsFeatureEnabled. I'm working with the shell team to determine whether there is a fix for this on their end.

Of the remaining ~100 ms spent in this method in the razor typing profile I'm looking at, 91 ms are coming from EnvironmentFeatureEnabled and IsVSServer methods in DefaultLSPEditorfeatureDetector. After talking with the razor team, these values can't really change during a VS session and thus can be cached.