Cache result from call to IVsFeatureFlags.IsFeatureEnabled
This call is very slow and shows up in the webtools razor typing perf test. There is an existing issue logged against the feature flags implementation, but it is prioritized low and guidance is to add your own cache around calls to it if you don't neeed them to respond to changes in the value.