Allow C# completion in middle of words.
- Prior to this C# completion would only ever be offered at start of word. Now we offer C# completion always to work around a racey:
When a user is doing 24/7 completion (just typing letters) HTML & C# only offer 24/7 completion at the beginning of a word. Meaning, completions will be provided at `|D` but not for `|Da` which brings us to an interesting cross-roads. Razor is currently designed with two language servers:
  1. HTML C#: Powers the HTML / C# experience
  2. Razor: Has all of the Razor understanding / powers the generated C# & HTML for a document
Because of this split, in the middle of completion requests it's possible for additional generated content (C# or HTML) to flow into the client. When additional content flows to the client we could mean to ask for completions at `|D` but in practice it'd ask C# for `|Da` (resulting in 0 completions). Therefore, to counteract this point-in-time design flaw we translate typing completion requests to explicit in order
to ensure that we still get completion results at `|Da`.
- Added tests to validate this new behavior.
- After changing C# behavior it revealed a separate unintended bug in the Razor completion engine. Basically another inherent race in how we do didChange requests. That's being addressed in future PRs (lots coming).

Part of #5017