Map SelectionRange for CodeActions requests (#5844)
* Map SelectionRange for CodeActions requests

If a SelectionRange is provided it's required that the range
should also be mapped correctly before sending off the request
for code actions. This context is used for cases where code action
ordering is ordered and the original range may not actually be
equal to the SelectionRange. Both are needed to correctly provide
the right code fixes and/or refactorings and in the right order
from Roslyn

* tests... maybe?

* Spaysing

* naming