Pull TagHelperDescriptor hashcode calculation into the tooling layer
The TagHelperDescriptorCache uses the hashcode of TagHelperDescriptors as a means of avoiding unnecesary deserialization of types. However, this means that the hashcode of the tag itself is used as an object equality check, and is a significant performance bottleneck in the compiler when tags are stored in a hashtable. This pulls the calculation of a cache ID into a dedicated method in the tooling layer, so we can remove the unnecessary elements from hashcode calculation for taghelpers for the general case.