Don't spin up 2 threads for every text edit (#6577)
* Use a batching work queue to prevent repeated work

At the very least this cuts the amount of work in half, as we hook up two events to this one method, so every text change would call this twice. Now the work queue ignores the first one for us.

Secondly, the timeout is very short, but still long enough that fast typing will not cause repeated work until there is a slight pause.

* Remove unused properties

Previously we reduced how often we would do a bunch of work, but it turns out a lot of that work wasn't ever used anyway. Seems to me like its better to just not do the work in the first place.

* Decrease timeout

* Remove debug code

* Doc