Un-genericize the aggregate completion provider
The idea behind this, when it was created, was to provide a generic way for there to be any number of completion providers, allow them to be requested in parallel, and then merge their results later. Unfortunately this doesn't work because in order to provide Razor tag helper completions we need to know what the Html completion contains.

Also there are only two providers, and I've no idea what other one we might introduce, so its all a bit of a waste of abstraction if you ask me. I suspect there was some theory that maybe the idea of an "aggregate" handler could be completely genericized to more than just completion too, but we can cross that bridge when if get to it.