Don't create tag helper completion items for Html tags
This code was creating tag helper completion items for any Html tag completion item that exists. Obviously this is misleading and incorrect, and only worked previously because there were never any Html tag completions passed in to this method. Instead we now treat Html completions and Razor completions that we've previously added, completely separately. This means adding another check in, but otherwise is fairly inconsequential.