Tweak how we decide we're "in" a HTML tag
This is the weirdest formatting bug I've ever fixed, because the fix is to make the code match the comment which describes it.

The previous check was waaayyy too broad.

It was saying that when we saw "    $$attr=value" we shouldn't format, because we're inside a HTML tag, which matches the method name, but the comment indicates we just want to ignore when inside HTML attributes. So that's what the new code does.

Surprisingly, none of the existing tests broke ¯\_(ツ)_/¯