Make FilePathNormalizer a static class
`FilePathNormalizer` is defined as a normal class and instances of it
are created. However, it's a general utility class with no subtypes
that doesn't have any fields. So, it can just be made static, which
cleans up a lot of code that was forced to always pass a
`FilePathNormalizer` instance.