Introduce LanguageSupport assembly (fixes #7886)
Microsoft.AspNetCore.Razor.Common is a bit *too* common for the
Razor compiler to reference. It's a bit of a dumping ground for things
that are important to tooling but not useful to the compiler, such as
VS telemetry and logging. When the compiler references Razor.Common,
there's an explosion of unnecessary dependencies.

This change introduces a new Microsoft.AspNetCore.Razor.LanguageSupport
assembly that includes types needed to use modern C# language features
in .NET Standard 2.0 and .NET Framework projects. This is referenced
by both the compiler and tooling. Razor.Common is now only referenced
by tooling and one lone compiler test project that uses a particular
helper from there.