Allow hover to defer to Roslyn if possible, for attributes
For Go To Definition, we only handled attributes on components, so it was safe to remove the code. For hover we also have code for attributes on MVC tag helpers which we don't want to break. Rather than trying to detect all of the cases, we can just defer to Roslyn when possible and we cover all bases.