Conditionally deploy Roslyn dependencies
Razor tooling deploys Roslyn dependencies to allow integration tests
and VS hive debugging to function against VS preview builds. However,
on local dev boxes, where VS IntPreview should be used for dogfooding,
this can result in hard-to-diagnose problems when some Roslyn
dependencies are loaded from the Razor tooling deployment and others
are loaded from the VS install.

This change only deploys Roslyn dependencies if the `IncludeRoslynDeps`
MSBuild property is set to true. This is needed for running integration
tests in CI, so the job has been updated to set that property.