Update tooling micro benchmark runner
When originally created, the tooling micro benchmarks lifted a small
bit of infrastructure from the aspnetcore repo to run BenchmarkDotNet.
This change removes that infrastructure and adds simpler logic to run
tests that is based on code from https://github.com/dotnet/performance.
In addition, BDN has been updated and a package reference has been added
to BenchmarkDotNet.Performance.Windows. This makes it easy to capture an
ETW profile on Windows.