Take version info out of DocumentContext, and move it to a new VersionedDocumentContext
A couple of bits of this aren't ideal, but are as good as we have for now.
Firstly, no covariant return types in .NET Framework mean the Identifier property can't be strongly typed as a VersionedTextDocumentIdentifier property without the new keyword. I think this turns out okay in practice though.
Secondly, it would be ideal to be able to call TryCreateCoreAsync without any state machine involved, but again the return types are against us.