Remove Microsoft.AspNetCore.Razor.Common assembly
With the introduction of Microsoft.AspNetCore.Razor.Utilities.Shared,
this assembly is no longer useful. The only types left in it were
in support telemetry, and those can easily be moved to
Microsoft.CodeAnalysis.Razor.Workspaces, which is already referenced by
most tooling projects. This will help reduce image loads in Visual
Studio.