Update all of the callers
Some of these changes only work on the assumption that the key is the project file, but its much easier to make all of these changes in one go, then start from there implementing a new system. All of the "bad" code in this commit is a call to `ProjectKey.From(string)`, so fixing the issues is a matter of finding those calls, and changing things.