Where only a project file is available, process all projects that match
This "fixes" the assumption that a project file represents a single project. At the moment it does, but the code at least doesn't have that assumption, and instead processes keys. If there is any opportunity in future to make these places smarter, and work on the most relevant project, then it should be easy enough to audit the calls to GetAllProjectKeys