Add RazorVersionedTextDocumentIdentifier (#8981)
Suggest commit by commit

﻿### Summary of the changes

With the addition of DocumentSymbol and ProjectContexts endpoints, we now need to flow VSProjectContext to more places. This change has the goal of doing that by:
Adding a new type for pairing a version and TextDocumentIdentifier. Since VersionedTextDocumentIdentifier doesn't have VSTextDocumentIdentifier as a base class, it would lose project context information. This allows the information to flow normally by manually adding our own pairing.

Add VSProjectContext to DocumentContext. Since we produce TextDocumentIdentifiers based on the context, this is needed to make sure we produce a VSTextDocumentIdentifier with the project context. Some places that call the factory don't currently have access to a project context, so they pass null. Most of these are code actions or things working with generated documents, which currently won't change based on framework selection. The only thing that can change is if they exist or not.

Change VersionedDocumentContext to return the new RazorVersionedTextDocumentIdentifier to be aligned with the above.