Add option to LanguageServerFeatureOptions to control generated file path behaviour
The C# extension in VS Code also generates file paths for generated files, and unless the logic in both sides match, nothing will work. Unfortunately we require more changes to give the C# extension the info it needs before it can change behaviour.

This also gives us an escape hatch to essentially turn off multi-target support if we need to, or a user needs to.