Update various code paths to allow for multiple virtual documents
This started as pseudocode to allow me to reason about the shape of the API I'd need, and ended up being something that is pretty close to done, and more importantly, something that should work fine with our current system. This could have been part of a subsequent PR, but I though it might help you all understand the API better.