Don't defer calculation of folding ranges, or error if they go wrong
The HandleCoreAsync method is called in a loop, and retries things, but also uses IEnumerables which meant that the actual calculation was deferred until the result was sent across the wire. I thought maybe this was causing issues with things operating on newer versions of a Razor document, with delegated responses coming from an older version. It also just didn't make sense with the retry logic.

The other change here is to stop a notification appearing in VS Code, that the user can't do anything about anyway, and switch to a log message.