Apply the same fix to UpdateHtml as we did to UpdateCSharp
One potential cause of incorrect folding ranges is if the Html content is "doubled" in the generated document. This is exactly the behaviour we saw with C# documents, I just completely missed applying it to Html.

As with C# documents, the "hack" part of this will be addressed in a future PR I'm preparing.