Attach test specific logging to the output window logging
Now that the output window uses a queue, to avoid threading issues, we were missing some logs as tests fail. This allows our integration tests to hook into the output window logging and ensure we capture it all directly into the test output.

What this means in practice is being able to download the one test run log file, and have it contain the full log for any failing test.