Generate nicer looking C#
Calling NormalizeWhitespace is the lazy way to not have to worry about trivia when using SyntaxFactory, but it also destroys any formatting the user might have done. Better to use the actual formatter to fix indentation and things, and just generate the code as a string.

Using syntax factory wasn't really helping too much.