Rework telemetry reporting to avoid allocations
This change reworks our telemetry reporting in the following ways:

1. Reduce allocations. I've removed usages of `ImmutableDictionary`
which has a lot of overhead and very poor lookup performance. In
addition, using a dictionary up front was a bit contrived because the
values are ultimately added to another dictionary on VS's
`TelemetryEvent` type. Instead of a dictionary, we now capture
properties in immutable arrays.
2. Refine the `ITelemetryReporter` interface. Over time, several
irrelevant methods have been added to the original slim design. Even
worse, some of the methods were only used by a particular
implementation and either threw or did nothing in other implementations,
which is definitely a smell.