Remove OmniSharp and update WebTools in tests
This turned out to be quite a rabbit hole to go down. This morning, I
noticed that an OmniSharp package was still referenced by our unit
tests. I decided to get rid of that but found that it was necessary
because our unit tests reference a very old version of the Web Tools
language server assemblies, which still depend on OmniSharp. So, I
needed to update those to the latest. However, updating Web Tools
required reworking all of the reflection code that interacts with
the Web Tools langauge server. Ultimately, I ended up introducing
a small reflection framework so that code interacting with Web Tools
types could be strongly typed.