Add nullability annotations to ProjectState
While adding nullability annotations to ProjectState, I decided to make
the ProjectState.ProjectWorkspaceState property non-nullable. It turns
out that a newly created ProjectState *might* have a null
ProjectWorkspaceState. It did not appear to me that this was useful,
so I added the invariant that a ProjectState will *always* have
a non-null ProjectWorkspaceState. If none is provided when a
ProjectState is created it'll get the ProjectWorkspaceState.Default
instance.