Remove IProjectSnapshotProjectEngineFactory
This is a much larger commit because it removes a fundamental service
that is used in a lot of places and replaces it with a simpler system.
Razor has a system for defining custom IProjectEngineFactory instances
whose jobs are to create RazorProjectEngines. This system was fairly
complex and clearly designed with extensibility in mind so that anyone
could export their own IProjectEngineFactory via MEF. However, this
extensibility isn't actually possible, since the IProjectEngineFactory
interface is internal. My guess is that this might have been a real
extensibility point many years ago, but was lost in some transition
to support a new platform (maybe ASP.NET Core? maybe LSP?).

There is now a simpler service called IProjectEngineFactoryProvider.
The job of this service is to locate the correct factory to use from
a static list of choices. There is likely further work that can be done
here, but this change removes much of the complexity.