Fully uninitialize the project if the intermediate output path changes
The previous code here was a great head start (thanks Phil!) but didn't quite do enough, because it only removed the project, it didn't tell the config file system to stop monitoring the old file. Additionally, it was a bit overcomplicated to pass the old key around, as the only reason the IntermediateOutputPath would appear in the "Before" snapshot is if it was changing.