Clear generated content on project remove
As soon as a project changes intermediate output path, we start sending updates for the new project. At the end of the publish though, things change back to the old one, which everything sees as a new project, but we've still got the previous generated content in memory on the server, so we send deltas when we should be sending the whole file at least once. This should help that!