Fix hang on solution close
There are scenarios on 17.10p2 where my recent changes to make
`FallbackProjectManager` use the dispatcher for project snapshot
manager changes can cause a dead lock. The problem occurs when
dynamic file info is updated on the dispatcher thread. Roslyn will then
call `GetDynamicFileInfoAsync(...)` on the dispatcher thread and block
on the result. `GetDynamicFileInfoAsync(...)` calls into
`FallbackProjectManager.DynamicFileAddedAsync(...)`, which attempts
to make changes to the project snapshot manager using the dispatcher.
However, because the dispatcher thread is blocked, it can never make
progress to newly queued task and hangs.

This is a small change to `FallbackProjectManager` to apply the
project snapshot manager changes directly if it's running on the
dispatcher thread. Otherwise, it rusn the change on the dispatcher.