Remove ProjectSnapshotManagerBase mocks in tests
This change makes strides toward removing mocks of ProjectSnapshotManagerBase. The biggest changes are in RazorProjectServiceTest, which has been rewritten to use real Razor objects and test outputs rather than mocking bits and pieces and verifying that members are called.

While working on RazorProjectServiceTest.cs, I noticed a method exposed for tests that was only ever called by tests, so I removed that.