Don't get IVsRunningDocumentTable on background thread
The `VisualStudioEditorDocumentManager` listens for running document
table events and was retrieving the `IVsRunningDocumentTable` in its
constructor. Because this is an `[ImportingConstructor]` it will
almost certain be called on the thread pool and will implicitly
marshal to the UI thread to perform a QI as part of retrieving the
service. This can cause a hang if the UI thread is tied up with other
work, such as when a solution is loaded.

This change avoids retrieving the running doc table service until its
needed and we are already running on the UI thread.