Don't use CancellationTokenSource for disposal
This change should address issues in ProjectWorkspaceStateGenerator where update cancellation was tied to the base IDisposable implementation in a complex way. This could cause exceptions to be thrown unnecessarily if the generator is disposed while an update is in flight. Now we just track a boolean field as we did in the past.