Use a named pipe to communicate projectinfo in vscode (#10521)
This changes from a file watcher approach to using named pipes. The flow for notifications is now:

Set up a handler that gets the name to connect to for a named pipe. This is generated as a GUID to avoid squatting. The Roslyn side will get this information through the razor/initialize message.
Roslyn sets up the server, Razor sets up a client.
All notifications go through that stream. MessagePack serialization is used for ProjectInfo still to serialize to bytes.
This will require a dual insertion and VS Code changes.