Re-rework LspEditorFeatureDetector
My previous work on LspEditorFeatureDetector introduced a significant regression by removing the code that checked project capabilities. When a razor document is opened, the LspEditorFeatureDetector is queried to see if the LSP editor is enabled (i.e. the user hasn't enabled the legacy editor instead), and to see whether the project that the document belongs to supports the LSP editor. In the case of .NET Framework projects, the LSP editor can't ever be used, and my prior change broke that. (Many thanks to @alexgav for fixing my mistake!)

Giving the situation a bit more thought, I've reworked LspEditorFeatureDetector again to simplify the code and fix a different issue where we would start checking project capabilities where we hadn't before.