Fix NRE when invoking completion in empty document (#10610)
* Fix NRE when invoking completion in empty document

In empty document completion context will have RazorDocumentSyntax as Owner. RazorDoucmentSyntax has null as the parent, which makes sense. However compiler thinks we can't have a null here, while obviously we do.

I'm adding a null check and a test for this case. We would probably further discuss the implications here. I tried to track down how we get null in a non-nullable field, but it's a bit confusing since the class gets generated.

* Move RazorDocumentSyntax special case inside switch per CR suggestion

* Fixing build