Add IProjectSnapshot.ContainsDocument to test document paths cheaply
There's a fair amount of code in Razor that uses `IProjectSnapshot.GetDocument(...)` to test for the presence of a document with a given file path in a project. However, `GetDocument` has the side effect of creating a document snapshot which ultimately may be thrown away.

This change adds an `IProjectSnapshot.ContainsDocument(...)` method that tests for the presence of a document by file path and returns true or false without creating a new document snapshot. In addition, I've cleaned up the `IProjectSnapshot` implementations a bit.