Fixing auto-insert inside of a C# statement. (#10952)
Current code was defining allowed auto-insert scenario too narrowly, with the main issue being that at the moment of the auto-insert the end tag for the newly typed tag is missing and if there is an enclosing/parent tag with the same name, the inner tag ends up "stealing" the end tag from that. We need to walk up the tree looking for the unclosed parent tag with the same name.