Fix request duration logging (#11198)
When first introduced there was an error in assumption of how histogram telemetry worked. This fixes this by correctly reporting a unique histogram even per lsp method name rather than the first one that comes in (which, as it turns out, was always initialize)

Also cleaned up the wording so it was a little easier to understand.