Add GenerateJsonFiles test class
This is is similar to the GenerateBaselines class. It updates the JSON files in the shared files folder. It is included in MS.ANC.Razor.ProjectEngineHost.Test rather than MS.ANC.Razor.Test.Common.Tooling for a couple of reason:

1. That's where most of the serialization tests live.
2. MS.ANC.Razor.Test.Common.Tooling is not a test project, so while the IDE will discover tests in that project, it won't run them.

Note that this change also updates the JSON because I actually ran the test. The JSON formatting is a bit different, which is where there are modifications.