Simplify ProjectSnapshotManager.NotifyListeners(...)
While there were excellent comments to describe the behavior of the notification queue, it was still a bit subtle to first peek the notification and only dequeue it when the listeners had been notified. This change introduces a bool field to make it clearer when notifications can be processed. In addition, this change introduces a catch around notification processing and logs any exceptions.