Remove RazorCompletionItem.Items property
Every RazorCompletionItem creates an instance of ItemCollection and exposes it via an Items property.
The Items property constructs the ItemCollection lazily, but that doesn't matter since it accessed immediately after each RazorCompletionItem constructor call to add a "description info" object (used to build completion description tooltips). No other data is stored in the ItemCollection.

This is an unnecessary amount of overhead to store a single object per RazorCompletionItem. Internally, ItemCollection is implemented with a ConcurrentDictionary, so its definitely not cheap!

The change removes the Items property altogether and provides a DescriptionInfo property that is set via the constructor.