Replace ISolutionQueryOperations in hover with dedicated service
When determining what projects a component or tag helper is available within, there is slightly different behavior in the Razor language server vs. co-hosting. In the language server, there is a special case to skip the misc files project which isn't bothered with in co-hosting. The special case is manifested via the different implementations of ISolutionQueryOperations.GetProjectsContainingDocument.

This change introduces a new service, IComponentAvailabilityService, that is used to determine what projects a component or tag helper is available within. The intention is that this service contains the behavior difference between the language server and co-hosting rather than relying on ISolutionQueryOperations, which will be going away with proper Razor solution snapshots.