Remove RazorCodeDocument.Create(...) calls from tooling code
These removes the remaining calls to RazorCodeDocument.Create(RazorSourceDocument) and RazorCodeDocument.Create(RazorSourceDocument, ImmutableArray<RazorSourceDocument>) from tooling code and tests. Each call has been updated to use the Create(...) overload that also takes options. (Currently, that overload allows null to be passed for the options, but that will change in a later commit.)