In VS, don't initialize misc files project with the workspace root path
The language server always starts up by loading the misc files project with every *.razor or *.cshtml file in the workspace root path. This *might* be helpful for VS Code (though I'm unsure about that), but in VS, it's clearly unnecessary overhead. The RazorProjectService will regularly receive new project information from the IRazorProjectInfoDriver, so this is really unnecessary overhead.