WorkspaceRootPathWatcher: Fix extension comparison
This fixes a bug introduced in a prior commit to properly verify file extensions when a file is renamed. To do this, I've added a couple of polyfill-style methods to a new PathUtilities class located in Microsoft.AspNetCore.Utilities.Shared. These are the same as Path.GetExtension and Path.HasExtension but they provide span-based overloads for non-.NET Core scenarios.