RazorCodeDocument: Remove SetFileKind
RazorCodeDocument already exposes ParserOptions.FileKind, so a separate mechanism for getting and setting the file kind is both unneeded and a potential source of bugs.

Most of the changes are in tooling tests, which have unfortunately proliferated a bit of an anti-pattern for constructing code documents through copy-paste. I'll take a look at cleaning that up in a follow-up pull request. This change tries to keep test clean up to a minimum.