ProjectSnapshotManager: Ensure that Open/CloseDocument always work
It turns out that CloseDocument can easily fail resulting in IsDocumentOpen to report true for documents that are not actually open. This is because the _openDocumentSet is only updated after updating the project with a SourceText or TextLoader. For example, if the project has already been removed, CloseDocument wouldn't update the _openDocumentSet. This change updates _openDocumentSet before updating the document text, which should by more correct.