Audit calls of SyntaxNode.DescendentNodesAndTokens
This change audits all callers of SyntaxNode.DescendantNodesAndTokens (and DescendantNodesAndTokensAndSelf) to determine if they should be doing something different. In each case, some analysis has been done to determine whether the caller is interested in 1) nodes, 2) tokens, or 3) both. In the case of 1 or 2, the caller is updated to call SyntaxNode.DescendantNodes (or DescendantNodesAndSelf) or SyntaxNode.DescendingTokens respectively.