Fix #11712
A bug introduced in 8c4d77cdfe4b12fe5acb0090cfeda6cebdcc68e0 causes RazorSourceDocuments for imports to be created with the wrong file paths. Instead of a pointing to a physical file path on disk, they contain a relative path, which causes incorrect code generation. To fix this, call the GetSourceAsync(...) helper for the import. That will populate the RazorSourceDocument with the correct paths.