Refactors classified span computation
Refactors the way classified spans are computed for Razor code documents intooling.

This change replaces the usage of legacy compiler APIs with a new `ClassifiedSpanVisitor`.
The new implementation is streamlined to only produce the information needed for the `GetLanguageKind()` API, improving performance and reducing dependencies.

This also includes changing the `ClassifiedSpanInternal` to `ClassifiedSpan` and removing the dependency on `Microsoft.AspNetCore.Razor.Language.Legacy`.