Bound Attributes: Use strongly-typed metadata over MetadataCollection
This is a biggish change that removes MetadataCollection from BoundAttributeDescriptor. Taking the place of metadata is a strongly-typed "MetadataObject" that might be different, depending on what the attribute is supposed to represent. Currently, the compiler uses different metadata objects for component type parameters and component properties.

While the metadata objects aren't as simple or flexible as MetadataCollection, they are faster and allocate list memory.