Make the test actually fail
We were inadvertently rejecting any attribute related diagnostics its seems. Partly because the whitespace before the attribute name was being considered as part of the attribute, and partly because each individual token of the attribute value was being considered. I suspect this broke with FUSE and with the node/token reversal. So that's fun.